Functional ITF twai_ops
